requirements.txt
anthropic
streamlit
python-dotenv
google-generativeai
pandas 
streamlit-option-menu 
streamlit-card
numpy
streamlit-lottie
transformers
tokenizers 